Refactor posts table
