+++ src/components/posts-table/posts-table.js
@@ -1,7 +1,6 @@
 /* eslint-disable camelcase */
 import React from 'react';
 import PropTypes from 'prop-types';
-import dayjs from 'dayjs';
 import {
   Container,
   Title,
@@ -11,35 +10,11 @@ import {
   TableLink,
   Column,
 } from './posts-table.styles';
-
-const getLinkTextToDisplay = (text, maxLength) => {
-  if (text.length > maxLength) {
-    const lengthOfElipsis = 3;
-    const shortenedText = text.slice(0, maxLength - lengthOfElipsis);
-    return `${shortenedText}...`;
-  }
-  return text;
-};
-
-const getTimeOfPost = (createdAtUtc) => {
-  const milliseconds = 1000;
-  const dateOfPost = new Date(createdAtUtc * milliseconds);
-  return dateOfPost;
-};
-
-const getTimeToDisplay = (createdAtUtc) => {
-  const dateOfPost = getTimeOfPost(createdAtUtc);
-  return dayjs(dateOfPost).format('hh:mmA');
-};
-
-const sortPostsByDate = (posts) => {
-  if (posts.length === 0) return posts;
-  return posts.sort((a, b) => {
-    const a_date = getTimeOfPost(a.data.created_utc);
-    const b_date = getTimeOfPost(b.data.created_utc);
-    return a_date.getMinutes() - b_date.getMinutes();
-  });
-};
+import {
+  getLinkTextToDisplay,
+  getTimeToDisplay,
+  sortPostsByDate,
+} from './posts-table.helpers';
 
 const PostsTable = ({ posts }) => {
   const lengthOfTitleColumn = 373;
